+++ landingzones/tests/test_permissions_ajax.py
@@ -156,3 +156,71 @@ class TestZoneStatusInfoRetrieveAjaxView(LandingzonesPermissionTestBase):
         self.set_site_read_only()
         self.assert_response(self.url, self.good_users, 200)
         self.assert_response(self.url, self.bad_users, 403)
+
+
+class TestZoneChecksumStatusRetrieveAjaxView(LandingzonesPermissionTestBase):
+    """Tests for ZoneChecksumStatusRetrieveAjaxView permissions"""
+
+    def setUp(self):
+        super().setUp()
+        self.investigation = self.import_isa_from_file(SHEET_PATH, self.project)
+        self.study = self.investigation.studies.first()
+        self.assay = self.study.assays.first()
+        zone = self.make_landing_zone(
+            title=ZONE_TITLE,
+            project=self.project,
+            user=self.user_contributor,  # NOTE: Zone owner = user_contributor
+            assay=self.assay,
+            description=ZONE_DESC,
+            status='ACTIVE',
+            configuration=None,
+            config_data={},
+        )
+        self.url = reverse(
+            'landingzones:ajax_irods_checksum',
+            kwargs={'landingzone': zone.sodar_uuid},
+        )
+        self.post_data = {'paths': []}
+        self.good_users = [
+            self.superuser,
+            self.user_owner_cat,
+            self.user_delegate_cat,
+            self.user_owner,
+            self.user_delegate,
+            self.user_contributor,  # Zone owner
+        ]
+        self.bad_users = [
+            self.user_contributor_cat,
+            self.user_guest_cat,
+            self.user_finder_cat,
+            self.user_guest,
+            self.user_no_roles,
+            self.anonymous,
+        ]
+
+    def test_get(self):
+        """Test ZoneStatusInfoRetrieveAjaxView GET"""
+        self.assert_response(self.url, self.good_users, 200, method='POST')
+        self.assert_response(self.url, self.bad_users, 403, method='POST')
+        self.project.set_public()
+        self.assert_response(self.url, self.no_role_users, 403, method='POST')
+
+    @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
+    def test_get_anon(self):
+        """Test GET with anonymous access"""
+        self.project.set_public()
+        self.assert_response(self.url, self.anonymous, 403, method='POST')
+
+    def test_get_archive(self):
+        """Test GET with archived project"""
+        self.project.set_archive()
+        self.assert_response(self.url, self.good_users, 200, method='POST')
+        self.assert_response(self.url, self.bad_users, 403, method='POST')
+        self.project.set_public()
+        self.assert_response(self.url, self.no_role_users, 403, method='POST')
+
+    def test_get_read_only(self):
+        """Test GET with site read-only mode"""
+        self.set_site_read_only()
+        self.assert_response(self.url, self.good_users, 200, method='POST')
+        self.assert_response(self.url, self.bad_users, 403, method='POST')